requirements.txt
scapy>=2.7.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
psutil>=6.0.0
pyyaml>=6.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
geoip2>=4.8.0
asyncpg>=0.31.0
python-jose[cryptography]>=3.5.0
passlib[bcrypt]>=1.7.4
bcrypt>=5.0.0
slowapi>=0.1.9
python-multipart>=0.0.22
APScheduler>=3.11.2
httpx>=0.28.1
